Use debian packages for cmake again
On bookworm, 'cmake' is new enough that we can just use it. On
bullseye, we can get a new-enough package from backports. By including
'cmake' in the build-deps stage, we don't need to install it
separately in all the later build stages that need it.

See https://github.com/neondatabase/neon/pull/2699, where switched to
downloading and building a specific version.