storcon: handle ongoing deletions gracefully
Problem

Pageserver returns 409 (Conflict) if any of the shards are already
deleting the timeline. This resulted in an error being propagated
out of the HTTP handler and to the client. It's an expected scenario
so we should handle it nicely.

Summary of Changes

Instead of returning an error on 409s, we now bubble the status code up
and let the HTTP handler code retry until it gets a 404 or times out.