pageserver: add metric for number of zeroed pages on rel extend
Problem

Filling the gap in with zeroes is annoying for sharded ingest.
We are not sure it even happens in reality.

Summary of Changes

Add one global counter which tracks how many such gap blocks
we filled on relation extends. We can add more metrics once
we understand the scope.