pageserver: fix test builds without testing feature
This was extremely frustating to deal with. Issue was that test builds
would fail without the "testing" flag.

We would like for unit tests to have the "pageserver_api/testing"
feature enabled. However, when I do that, the non-test code does
not know how to handle the `NeonWalRecord::Test` variant.

Eventually, I settled on a compromise: any unit tests that require
the testing features of `NeonWalRecord` run only when the "testing"
feature is enabled on the "pagserver" crate. In CI we already enable
the "testing" feature when running unit tests for the "pageserver"
crate.