Coalesce NULL LFC stats values to 0
The Neon extension can work without the LFC enabled. When we query the
LFC stats while it is disabled, they return NULL. This is fine for most
use cases, but isn't fine for sql_exporter, a Prometheus exporter.
sql_exporter tries to create a double precision floating point value out
of a NULL value, which it logs as a failure and will subsequently not
return data for that metric.

Let's take a simple approach and coalesce NULL values to 0, which should
be easily understandable because since the LFC is disabled, there would
obviously be 0 hits, misses, etc.

Signed-off-by: Tristan Partin <tristan@neon.tech>