pageserver: cut over to batching in interpreted wal records
This is a bigger bite than I would have liked, but this stuff is
tightly coupled.

The main goal of this commit is to use the `SerializedValueBatch`
in `InterpretedWalRecord`. This requires that `DatadirModification`
maintains a `SerializedValueBatch` and extends it as new WAL records
come in (to avoid flushing to disk on every record). In turn, this
cascaded into a number of modifications to `DatadirModification`
1. Replace `pending_data_pages` and `pending_zero_data_pages` with
`pending_data_batch`.
2. Removal of `pending_zero_data_pages` and its cousin
`on_wal_record_end`
3. Rename `pending_bytes` to `pending_metadata_bytes` since this is what
it tracks now.
4. Adapting of various utility methods like `len`,
`approx_pending_bytes` and `has_dirty_data_pages`.

Removal of `pending_zero_data_pages` and the optimisation associated
with it ((1) and (2)) deserves more detail.

Previously all zero data pages went through `pending_zero_data_pages`.
We wrote zero data pages when filling gaps caused by relation extension
(case A) and when handling special wal records (case B). If it happened
that the same WAL record contained a non zero write for an entry in
`pending_zero_data_pages` we skipped the zero write.

Case A: We handle relation extensions differently now. Once the batch
grows large enough, we identify the gaps and fill the gaps when handling
the entire wal record. Hence, the optimisation is not required
anymore.

Case B: When the handling of a special record needs to zero out a key,
it just adds that to the current batch. I inspected the code, and I
don't think the optimisation kicked in here.