tests: Increase timeout in test_create_churn_during_restart
This test was seen to be flaky, e.g. at
https://neon-github-public-dev.s3.amazonaws.com/reports/pr-9457/11804246485/index.html#suites/ec4311502db344eee91f1354e9dc839b/982bd121ea698414/.
If I _reduce_ the timeout from 10s to 8s on my laptop, it reliably
hits that timeout and fails. That suggests that the test is pretty
close to the edge even when it passes. Let's bump up the timeout to 30
s to make it more robust.

See also https://github.com/neondatabase/neon/issues/9730, although
the error message is different there.